+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `onStakePoolChange` and `onValidatorListChange`
+- Export account schema types and classes
 
 ## [0.3.0] - 2022-04-28