+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+- Export error types
+
 ## [0.4.2] - 2022-05-16
 
 ### Changed
@@ -17,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2022-05-10
 
 ### Changed
-- `depositStake` and `depositStakeTransactions` now take an additional `amount` argument that can be used to split an active stake account before depositing. 
+- `depositStake` and `depositStakeTransactions` now take an additional `amount` argument that can be used to split an active stake account before depositing.
 
 ### Added
 - `onStakePoolChange` and `onValidatorListChange`