+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.4.3] - 2022-09-16
 
-### Fixed
+### Changed
 - Export error types
+- Export `decodeStakePool` and `decodeValidatorList`
 
 ## [0.4.2] - 2022-05-16